refactor: extract computation of performance rates to separate function
